+++ comb_spec_searcher/strategies/rule.py
@@ -26,6 +26,7 @@ from typing import (
     cast,
 )
 
+from logzero import logger
 from sympy import Eq, Function
 
 from comb_spec_searcher.combinatorial_class import CombinatorialClass
@@ -365,9 +366,17 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
         assert isinstance(strategy, Strategy)
         comb_class = CombinatorialClass.from_dict(d.pop("comb_class"))
         comb_class = cast(CombinatorialClassType, comb_class)
-        d.pop("children")
+        saved_children = tuple(map(CombinatorialClass.from_dict, d.pop("children")))
+        rule = cls(strategy, comb_class)
+        if rule.children != saved_children:
+            msg = "Children built from strategy different from saved children\n"
+            msg += f"{rule}\n"
+            msg += "The saved children are:\n"
+            for child in saved_children:
+                msg += f"{child}\n"
+            logger.warning(msg)
         assert not d
-        return cls(strategy, comb_class)
+        return rule
 
     @property
     def strategy(self) -> "Strategy[CombinatorialClassType, CombinatorialObjectType]":